Import a dataset from monolix (based on an imported model)

Import a dataset from monolix (based on an imported model)

monolixDataImport(ui, data, na.strings = c("NA", "."))

Arguments

ui

A rxode2 ui model imported from monolix

data

dataset to convert to nlmixr2 format; if the dataset is missing, load from the mlxtran specification

na.strings

a character vector of strings which are to be interpreted as NA values. Blank fields are also considered to be missing values in logical, integer, numeric and complex fields. Note that the test happens after white space is stripped from the input (if enabled), so na.strings values may need their own white space stripped in advance.

Value

Dataset appropriate for using with the rxode2 model for simulations

Author

Matthew L. Fidler

Examples

# First load in the model; in this case the theo model
# This is modified from the Monolix demos by saving the model
# File as a text file (hence you can access without model library)
# setup.
#
# This example is also included in the monolix2rx package, so
# you refer to the location with `system.file()`:

pkgTheo <- system.file("theo", package="monolix2rx")

mod <- monolix2rx(file.path(pkgTheo, "theophylline_project.mlxtran"))
#> ℹ updating model values to final parameter estimates
#> ℹ done
#> ℹ reading run info (# obs, doses, Monolix Version, etc) from summary.txt
#> ℹ done
#> ℹ reading covariance from FisherInformation/covarianceEstimatesLin.txt
#> ℹ done
#> ℹ imported monolix and translated to rxode2 compatible data ($monolixData)
#> ℹ imported monolix ETAS (_SAEM) imported to rxode2 compatible data ($etaData)
#> ℹ imported monolix pred/ipred data to compare ($predIpredData)
#>  
#>  
#> using C compiler: ‘gcc (Ubuntu 11.4.0-1ubuntu1~22.04) 11.4.0’
#> ℹ solving ipred problem
#> ℹ done
#> ℹ solving pred problem
#> ℹ done

# read in monolix dataset

dat <- read.table(file.path(pkgTheo, "data", "theophylline_data.txt"),na=".", header=TRUE)

monolixDataImport(mod, dat)
#>     id  amt  time    dv WEIGHT SEX  cmt admd
#> 1    1 4.02  0.00    NA   79.6  NA <NA>   NA
#> 2    1   NA  0.25  2.84   79.6  NA <NA>   NA
#> 3    1   NA  0.57  6.57   79.6  NA <NA>   NA
#> 4    1   NA  1.12 10.50   79.6  NA <NA>   NA
#> 5    1   NA  2.02  9.66   79.6  NA <NA>   NA
#> 6    1   NA  3.82  8.58   79.6  NA <NA>   NA
#> 7    1   NA  5.10  8.36   79.6  NA <NA>   NA
#> 8    1   NA  7.03  7.47   79.6  NA <NA>   NA
#> 9    1   NA  9.05  6.89   79.6  NA <NA>   NA
#> 10   1   NA 12.12  5.94   79.6  NA <NA>   NA
#> 11   1   NA 24.37  3.28   79.6  NA <NA>   NA
#> 12   2 4.40  0.00    NA   72.4  NA <NA>   NA
#> 13   2   NA  0.27  1.72   72.4  NA <NA>   NA
#> 14   2   NA  0.52  7.91   72.4  NA <NA>   NA
#> 15   2   NA  1.00  8.31   72.4  NA <NA>   NA
#> 16   2   NA  1.92  8.33   72.4  NA <NA>   NA
#> 17   2   NA  3.50  6.85   72.4  NA <NA>   NA
#> 18   2   NA  5.02  6.08   72.4  NA <NA>   NA
#> 19   2   NA  7.03  5.40   72.4  NA <NA>   NA
#> 20   2   NA  9.00  4.55   72.4  NA <NA>   NA
#> 21   2   NA 12.00  3.01   72.4  NA <NA>   NA
#> 22   2   NA 24.30  0.90   72.4  NA <NA>   NA
#> 23   3 4.53  0.00    NA   70.5  NA <NA>   NA
#> 24   3   NA  0.27  4.40   70.5  NA <NA>   NA
#> 25   3   NA  0.58  6.90   70.5  NA <NA>   NA
#> 26   3   NA  1.02  8.20   70.5  NA <NA>   NA
#> 27   3   NA  2.02  7.80   70.5  NA <NA>   NA
#> 28   3   NA  3.62  7.50   70.5  NA <NA>   NA
#> 29   3   NA  5.08  6.20   70.5  NA <NA>   NA
#> 30   3   NA  7.07  5.30   70.5  NA <NA>   NA
#> 31   3   NA  9.00  4.90   70.5  NA <NA>   NA
#> 32   3   NA 12.15  3.70   70.5  NA <NA>   NA
#> 33   3   NA 24.17  1.05   70.5  NA <NA>   NA
#> 34   4 4.40  0.00    NA   72.7  NA <NA>   NA
#> 35   4   NA  0.35  1.89   72.7  NA <NA>   NA
#> 36   4   NA  0.60  4.60   72.7  NA <NA>   NA
#> 37   4   NA  1.07  8.60   72.7  NA <NA>   NA
#> 38   4   NA  2.13  8.38   72.7  NA <NA>   NA
#> 39   4   NA  3.50  7.54   72.7  NA <NA>   NA
#> 40   4   NA  5.02  6.88   72.7  NA <NA>   NA
#> 41   4   NA  7.02  5.78   72.7  NA <NA>   NA
#> 42   4   NA  9.02  5.33   72.7  NA <NA>   NA
#> 43   4   NA 11.98  4.19   72.7  NA <NA>   NA
#> 44   4   NA 24.65  1.15   72.7  NA <NA>   NA
#> 45   5 5.86  0.00    NA   54.6  NA <NA>   NA
#> 46   5   NA  0.30  2.02   54.6  NA <NA>   NA
#> 47   5   NA  0.52  5.63   54.6  NA <NA>   NA
#> 48   5   NA  1.00 11.40   54.6  NA <NA>   NA
#> 49   5   NA  2.02  9.33   54.6  NA <NA>   NA
#> 50   5   NA  3.50  8.74   54.6  NA <NA>   NA
#> 51   5   NA  5.02  7.56   54.6  NA <NA>   NA
#> 52   5   NA  7.02  7.09   54.6  NA <NA>   NA
#> 53   5   NA  9.10  5.90   54.6  NA <NA>   NA
#> 54   5   NA 12.00  4.37   54.6  NA <NA>   NA
#> 55   5   NA 24.35  1.57   54.6  NA <NA>   NA
#> 56   6 4.00  0.00    NA   80.0  NA <NA>   NA
#> 57   6   NA  0.27  1.29   80.0  NA <NA>   NA
#> 58   6   NA  0.58  3.08   80.0  NA <NA>   NA
#> 59   6   NA  1.15  6.44   80.0  NA <NA>   NA
#> 60   6   NA  2.03  6.32   80.0  NA <NA>   NA
#> 61   6   NA  3.57  5.53   80.0  NA <NA>   NA
#> 62   6   NA  5.00  4.94   80.0  NA <NA>   NA
#> 63   6   NA  7.00  4.02   80.0  NA <NA>   NA
#> 64   6   NA  9.22  3.46   80.0  NA <NA>   NA
#> 65   6   NA 12.10  2.78   80.0  NA <NA>   NA
#> 66   6   NA 23.85  0.92   80.0  NA <NA>   NA
#> 67   7 4.95  0.00    NA   64.6  NA <NA>   NA
#> 68   7   NA  0.25  0.85   64.6  NA <NA>   NA
#> 69   7   NA  0.50  2.35   64.6  NA <NA>   NA
#> 70   7   NA  1.02  5.02   64.6  NA <NA>   NA
#> 71   7   NA  2.02  6.58   64.6  NA <NA>   NA
#> 72   7   NA  3.48  7.09   64.6  NA <NA>   NA
#> 73   7   NA  5.00  6.66   64.6  NA <NA>   NA
#> 74   7   NA  6.98  5.25   64.6  NA <NA>   NA
#> 75   7   NA  9.00  4.39   64.6  NA <NA>   NA
#> 76   7   NA 12.05  3.53   64.6  NA <NA>   NA
#> 77   7   NA 24.22  1.15   64.6  NA <NA>   NA
#> 78   8 4.53  0.00    NA   70.5  NA <NA>   NA
#> 79   8   NA  0.25  3.05   70.5  NA <NA>   NA
#> 80   8   NA  0.52  3.05   70.5  NA <NA>   NA
#> 81   8   NA  0.98  7.31   70.5  NA <NA>   NA
#> 82   8   NA  2.02  7.56   70.5  NA <NA>   NA
#> 83   8   NA  3.53  6.59   70.5  NA <NA>   NA
#> 84   8   NA  5.05  5.88   70.5  NA <NA>   NA
#> 85   8   NA  7.15  4.73   70.5  NA <NA>   NA
#> 86   8   NA  9.07  4.57   70.5  NA <NA>   NA
#> 87   8   NA 12.10  3.00   70.5  NA <NA>   NA
#> 88   8   NA 24.12  1.25   70.5  NA <NA>   NA
#> 89   9 3.10  0.00    NA   86.4  NA <NA>   NA
#> 90   9   NA  0.30  7.37   86.4  NA <NA>   NA
#> 91   9   NA  0.63  9.03   86.4  NA <NA>   NA
#> 92   9   NA  1.05  7.14   86.4  NA <NA>   NA
#> 93   9   NA  2.02  6.33   86.4  NA <NA>   NA
#> 94   9   NA  3.53  5.66   86.4  NA <NA>   NA
#> 95   9   NA  5.02  5.67   86.4  NA <NA>   NA
#> 96   9   NA  7.17  4.24   86.4  NA <NA>   NA
#> 97   9   NA  8.80  4.11   86.4  NA <NA>   NA
#> 98   9   NA 11.60  3.16   86.4  NA <NA>   NA
#> 99   9   NA 24.43  1.12   86.4  NA <NA>   NA
#> 100 10 5.50  0.00    NA   58.2  NA <NA>   NA
#> 101 10   NA  0.37  2.89   58.2  NA <NA>   NA
#> 102 10   NA  0.77  5.22   58.2  NA <NA>   NA
#> 103 10   NA  1.02  6.41   58.2  NA <NA>   NA
#> 104 10   NA  2.05  7.83   58.2  NA <NA>   NA
#> 105 10   NA  3.55 10.21   58.2  NA <NA>   NA
#> 106 10   NA  5.05  9.18   58.2  NA <NA>   NA
#> 107 10   NA  7.08  8.02   58.2  NA <NA>   NA
#> 108 10   NA  9.38  7.14   58.2  NA <NA>   NA
#> 109 10   NA 12.10  5.68   58.2  NA <NA>   NA
#> 110 10   NA 23.70  2.42   58.2  NA <NA>   NA
#> 111 11 4.92  0.00    NA   65.0  NA <NA>   NA
#> 112 11   NA  0.25  4.86   65.0  NA <NA>   NA
#> 113 11   NA  0.50  7.24   65.0  NA <NA>   NA
#> 114 11   NA  0.98  8.00   65.0  NA <NA>   NA
#> 115 11   NA  1.98  6.81   65.0  NA <NA>   NA
#> 116 11   NA  3.60  5.87   65.0  NA <NA>   NA
#> 117 11   NA  5.02  5.22   65.0  NA <NA>   NA
#> 118 11   NA  7.03  4.45   65.0  NA <NA>   NA
#> 119 11   NA  9.03  3.62   65.0  NA <NA>   NA
#> 120 11   NA 12.12  2.69   65.0  NA <NA>   NA
#> 121 11   NA 24.08  0.86   65.0  NA <NA>   NA
#> 122 12 5.30  0.00    NA   60.5  NA <NA>   NA
#> 123 12   NA  0.25  1.25   60.5  NA <NA>   NA
#> 124 12   NA  0.50  3.96   60.5  NA <NA>   NA
#> 125 12   NA  1.00  7.82   60.5  NA <NA>   NA
#> 126 12   NA  2.00  9.72   60.5  NA <NA>   NA
#> 127 12   NA  3.52  9.75   60.5  NA <NA>   NA
#> 128 12   NA  5.07  8.57   60.5  NA <NA>   NA
#> 129 12   NA  7.07  6.59   60.5  NA <NA>   NA
#> 130 12   NA  9.03  6.11   60.5  NA <NA>   NA
#> 131 12   NA 12.05  4.57   60.5  NA <NA>   NA
#> 132 12   NA 24.15  1.17   60.5  NA <NA>   NA